🎨 친구 컨트롤러 패키지 수정 및 API 명세 보강
- dto 내 패키지를 request/response로 분리
- 각 Request/Response DTO 객체에 ApiModelProperty 애노테이션으로 필드 설명 추가
