+++ src/main/java/com/nexters/keyme/friend/presentation/controller/FriendController.java
@@ -3,9 +3,9 @@ package com.nexters.keyme.friend.presentation.controller;
 import com.nexters.keyme.auth.dto.UserInfo;
 import com.nexters.keyme.auth.resolver.RequestUser;
 import com.nexters.keyme.common.dto.ApiResponse;
-import com.nexters.keyme.friend.presentation.dto.FriendAcceptRequest;
-import com.nexters.keyme.friend.presentation.dto.FriendAddRequest;
-import com.nexters.keyme.friend.presentation.dto.FriendResponse;
+import com.nexters.keyme.friend.presentation.dto.request.FriendAcceptRequest;
+import com.nexters.keyme.friend.presentation.dto.request.FriendAddRequest;
+import com.nexters.keyme.friend.presentation.dto.response.FriendResponse;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
