--- src/main/java/com/nexters/keyme/friend/presentation/dto/FriendResponse.java
@@ -1,12 +1,0 @@
-package com.nexters.keyme.friend.presentation.dto;
-
-import lombok.*;
-
-@Getter
-@RequiredArgsConstructor
-@Builder
-public class FriendResponse {
-    private final Long memberId;
-    private final String nickname;
-    private final String thumbnailUrl;
-}
