+++ src/main/java/com/nexters/keyme/friend/presentation/dto/request/FriendAcceptRequest.java
@@ -1,5 +1,6 @@
-package com.nexters.keyme.friend.presentation.dto;
+package com.nexters.keyme.friend.presentation.dto.request;
 
+import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,6 +9,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FriendAcceptRequest {
+    @ApiModelProperty(value="수락할 친구 ID", example = "22")
     private Long acceptFriendId;
+    @ApiModelProperty(value="수락 여부", example = "true")
     private boolean isAccept;
 }
